modification DAO

+++ src/org/eni/encheres/bo/Utilisateur.java
@@ -11,11 +11,11 @@ public class Utilisateur {
 	private String codePostal;
 	private String ville;
 	private String motDePasse;
-	private float credit;
+	private int credit;
 	private Boolean administrateur;
 	
 	public Utilisateur(int noUtilisateur, String pseudo, String nom, String prenom, String email, String telephone,
-			String rue, String codePostal, String ville, String motDePasse, float credit, Boolean administrateur) {
+			String rue, String codePostal, String ville, String motDePasse, int credit, Boolean administrateur) {
 		super();
 		this.noUtilisateur = noUtilisateur;
 		this.pseudo = pseudo;
@@ -125,11 +125,11 @@ public class Utilisateur {
 		this.motDePasse = motDePasse;
 	}
 
-	public float getCredit() {
+	public int getCredit() {
 		return credit;
 	}
 
-	public void setCredit(float credit) {
+	public void setCredit(int credit) {
 		this.credit = credit;
 	}
